extract websocket tests in separate module
